% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_upper_addin.R
\name{to_upper_addin}
\alias{to_upper_addin}
\title{Transform selected text to uppercase}
\usage{
to_upper_addin()
}
\description{
\code{to_upper_addin} uses the rstudioapi to get the selected text
then replace it with the uppercase version
}
