% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_trailing_whitespace_addin.R
\name{trim_trailing_whitespace_addin}
\alias{trim_trailing_whitespace_addin}
\title{Trim trailing whitespace}
\usage{
trim_trailing_whitespace_addin()
}
\description{
\code{trim_trailing_whitespace_addin} uses the rstudioapi to get the selected text
then remove all trailing whitespace
}
